Add basic chatbot states
Each chatbot on a given server will have a "current state" and a
"desired state". The state determines which modules should be loaded and
which rooms should be joined. The implementation is minimal for now, but
will be expanded soon. This will become useful when auto configuration
reloading is implemented.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>