Migrate chatbot states to be CLI-only, and add structure to CLI
The CLI program was the only thing using the chatbot states, and I think
that's how it should stay. It's merely a wrapper for the actual chatbot
stuff, and since its "desired state" switching is part of the hot config
reloading (which is dictated by the CLI), I've moved it to the CLI only.
Much less spaghetti.

Beyond that, we have actual objects for the CLI, so there's some
idiomatic python going on, and less awful ad-hoc configuration and setup
in the "main" method.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>