Update module handling and loading to be hotload-friendly
Config autoreload is coming along nicely. These changes clean up a few
mistakes in the typing of modules (moving from a list to dict
internally) and organization of module loading. Chatbots are now
responsible for the loading of their modules, rather than the driver,
which cleans things up a lot.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>