Finish off config reloading
Things are still kind of ad-hoc right now, especially with file
watching. Usually when files get updated via a modern editor, instead of
modifying the file, it deletes it, writes to a backup, and copies the
backup over instead, which isn't as simple to detect.

Either way, things are still a little hacky but it's at least usable. I
hope to put examples in there soon so people can implement their own
bots :)

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>