Refactor and apply changes to jayk.cli.module
This was a bigger one, so there are going to be a lot of lines changed.
The only main thing I'm concerned about is that the JaykMeta metaclass
was accepting a **kwargs argument which wasn't being used. I don't think
it needed to be in there, but if it breaks, we'll know the culprit.

Lots of documentation in this one, too.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>