Add first revision of TODO list for v1.0 release.
I've been teetering on the edge of a "first release" for a while and I
want to get features that we want in that release to be finalized. A
huge part of the TODO is the documentation that needs to be done. I want
to focus on this for now and implement features later (since most
features are working and implemented).

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>